+++ insts/std/MMOVE/MMOVE.h
@@ -10,7 +10,7 @@ public:
     MOVE();
     virtual ~MOVE();
 protected:
-    virtual int localInit(float *, int);
+    virtual int localInit(double *, int);
     virtual int finishInit(double, double *);
     virtual int updatePosition(int);
     virtual void get_tap(int, int, int, int);
