+++ insts/std/MMOVE/MPLACE.cpp
@@ -36,7 +36,7 @@ PLACE::~PLACE()
 {
 }
 
-int PLACE::localInit(float p[], int n_args)
+int PLACE::localInit(double p[], int n_args)
 {
     if (n_args < 8) {
         die(name(), "Wrong number of args.");
